fix start cmd for WSL
Assuming this was derived from the One-click installers section in the readme file in the oobabooga repo, where it states `Run the start_linux.sh, start_windows.bat, start_macos.sh, or start_wsl.bat script depending on your OS.`, the `start_wsl.bat` is actually meant to run from a Windows command prompt or PowerShell. In the context of this ooba implementation, it should just treat WSL as Linux.