Some fixes on paths on windows and finding open sockets using async
path's on windows doesn't contain "/" but "\". I used os.path.dirname/basename to get folder path and file path.

Replaced the entire get_open_ports.py so it finds the open ports... Also made it async, because I don't want to wait for a few days for it to finish...